MFC r312063:
Ensure that the buffer length and the length provided in the IPv4
header match when using a raw socket to send IPv4 packets and
providing the header. If they don't match, let send return -1
and set errno to EINVAL.

Before this patch is was only enforced that the length in the header
is not larger then the buffer length.

PR:			212283
Reviewed by:		ae, gnn
Differential Revision:	https://reviews.freebsd.org/D9161


git-svn-id: https://svn.freebsd.org/base/stable/11@319392 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f