MFC r314155:
TCP window updates are only sent if the window can be increased by at
least 2 * MSS. However, if the receive buffer size is small, this might
be impossible. Add back a criterion to send a TCP window update if
the window can be increased by at least half of the receive buffer size.
This condition was removed in r242252. This patch simply brings it back.

PR:			211003
Reviewed by:		gnn
Differential Revision:	https://reviews.freebsd.org/D9475


git-svn-id: https://svn.freebsd.org/base/stable/11@319397 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f