MFC r317208:
Syncoockies can be used in combination with the syncache. If the cache
overflows, syncookies are used.
This patch restricts the usage of syncookies in this case: accept
syncookies only if there was an overflow of the syncache recently.
This mitigates a problem reported in PR217637, where is syncookie was
accepted without any recent drops.
Thanks to glebius@ for suggesting an improvement.

PR:			217637
Reviewed by:		gnn, glebius
Differential Revision:	https://reviews.freebsd.org/D10272


git-svn-id: https://svn.freebsd.org/base/stable/11@319400 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f