Remove an unnecessary field from struct blist.  (The comment describing what this field represented was also inaccurate.)  Suggested by: kib
In r178792, blist_create() grew a malloc flag, allowing M_NOWAIT to be
specified.  However, blist_create() was not modified to handle the
possibility that a malloc() call failed.  Address this omission.

Increase the width of the local variable "radix" to 64 bits.  (This
matches the width of the corresponding field in struct blist.)

Reviewed by:	kib
MFC after:	6 weeks


git-svn-id: https://svn.freebsd.org/base/head@319793 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f