Change blist_alloc()'s allocation policy from first-fit to next-fit so that disk writes are more likely to be sequential.  This change is beneficial on both the solid state and mechanical disks that I've tested.  (A similar change in allocation policy was made by DragonFly BSD in 2013 to speed up Poudriere with "stressful memory parameters".)
Increase the width of blst_meta_alloc()'s parameter "skip" and the local
variables whose values are derived from it to 64 bits.  (This matches the
width of the field "skip" that is stored in the structure "blist" and
passed to blst_meta_alloc().)

Eliminate a pointless check for a NULL blist_t.

Simplify blst_meta_alloc()'s handling of the ALL-FREE case.

Address nearby style errors.

Reviewed by:	kib, markj
MFC after:	5 weeks
Differential Revision:	https://reviews.freebsd.org/D11247


git-svn-id: https://svn.freebsd.org/base/head@320077 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f