Add iic_recover_bus(), a helper function that can be used by any i2c driver which is able to manipulate the clock and data lines directly.
When an i2c bus is hung by a slave device stuck in the middle of a
transaction that didn't complete properly, this function manipulates the
clock and data lines in a sequence known to reliably reset slave devices.
The most common cause of a hung i2c bus is a system reboot in the middle of
an i2c transfer (so it doesnt' happen often, but now there is a way other
than power cycling to recover from it).


git-svn-id: https://svn.freebsd.org/base/head@320461 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f